Fixes output file assertions in tests
According to the youtube-dl documentation, the output option
allows to set the template of the resulting file.
This means that there can not be a hard expectation on the file
extension when there is no preference given.

Since mp4 seems to be the default preferred type I set those
assertions to "nope.mp4".